+++ bridge/bridge.go
@@ -7,6 +7,7 @@
 package bridge
 
 import (
+	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -15,11 +16,18 @@ import (
 	"os/signal"
 	"runtime"
 	"strings"
+	"sync"
 	"syscall"
 	"time"
 
+	"github.com/lib/pq"
 	"github.com/mattn/go-sqlite3"
 	"github.com/rs/zerolog"
+	deflog "github.com/rs/zerolog/log"
+	"go.mau.fi/util/configupgrade"
+	"go.mau.fi/util/dbutil"
+	_ "go.mau.fi/util/dbutil/litestream"
+	"go.mau.fi/util/exzerolog"
 	"gopkg.in/yaml.v3"
 	flag "maunium.net/go/mauflag"
 	"maunium.net/go/maulogger/v2"
@@ -28,12 +36,10 @@ import (
 	"maunium.net/go/mautrix"
 	"maunium.net/go/mautrix/appservice"
 	"maunium.net/go/mautrix/bridge/bridgeconfig"
+	"maunium.net/go/mautrix/bridge/status"
 	"maunium.net/go/mautrix/event"
 	"maunium.net/go/mautrix/id"
 	"maunium.net/go/mautrix/sqlstatestore"
-	"maunium.net/go/mautrix/util/configupgrade"
-	"maunium.net/go/mautrix/util/dbutil"
-	_ "maunium.net/go/mautrix/util/dbutil/litestream"
 )
 
 var configPath = flag.MakeFull("c", "config", "The path to your config file.", "config.yaml").String()
@@ -41,6 +47,7 @@ var dontSaveConfig = flag.MakeFull("n", "no-update", "Don't save updated config 
 var registrationPath = flag.MakeFull("r", "registration", "The path where to save the appservice registration.", "registration.yaml").String()
 var generateRegistration = flag.MakeFull("g", "generate-registration", "Generate registration and quit.", "false").Bool()
 var version = flag.MakeFull("v", "version", "View bridge version and quit.", "false").Bool()
+var versionJSON = flag.Make().LongKey("version-json").Usage("Print a JSON object representing the bridge version and quit.").Default("false").Bool()
 var ignoreUnsupportedDatabase = flag.Make().LongKey("ignore-unsupported-database").Usage("Run even if the database schema is too new").Default("false").Bool()
 var ignoreForeignTables = flag.Make().LongKey("ignore-foreign-tables").Usage("Run even if the database contains tables from other programs (like Synapse)").Default("false").Bool()
 var wantHelp, _ = flag.MakeHelpFlag()
@@ -54,14 +61,14 @@ type Portal interface {
 	MainIntent() *appservice.IntentAPI
 
 	ReceiveMatrixEvent(user User, evt *event.Event)
-	UpdateBridgeInfo()
+	UpdateBridgeInfo(ctx context.Context)
 }
 
 type MembershipHandlingPortal interface {
 	Portal
-	HandleMatrixLeave(sender User)
-	HandleMatrixKick(sender User, ghost Ghost)
-	HandleMatrixInvite(sender User, ghost Ghost)
+	HandleMatrixLeave(sender User, evt *event.Event)
+	HandleMatrixKick(sender User, ghost Ghost, evt *event.Event)
+	HandleMatrixInvite(sender User, ghost Ghost, evt *event.Event)
 }
 
 type ReadReceiptHandlingPortal interface {
@@ -97,6 +104,7 @@ type User interface {
 type DoublePuppet interface {
 	CustomIntent() *appservice.IntentAPI
 	SwitchCustomMXID(accessToken string, userID id.UserID) error
+	ClearCustomMXID()
 }
 
 type Ghost interface {
@@ -127,6 +135,11 @@ type ChildOverride interface {
 	CreatePrivatePortal(id.RoomID, User, Ghost)
 }
 
+type ConfigValidatingBridge interface {
+	ChildOverride
+	ValidateConfig() error
+}
+
 type FlagHandlingBridge interface {
 	ChildOverride
 	HandleFlags() bool
@@ -135,6 +148,11 @@ type FlagHandlingBridge interface {
 type PreInitableBridge interface {
 	ChildOverride
 	PreInit()
+}
+
+type WebsocketStartingBridge interface {
+	ChildOverride
+	OnWebsocketConnect()
 }
 
 type CSFeatureRequirer interface {
@@ -156,8 +174,12 @@ type Bridge struct {
 	VersionDesc      string
 	LinkifiedVersion string
 	BuildTime        string
+	commit           string
+	baseVersion      string
 
 	PublicHSAddress *url.URL
+
+	DoublePuppet *doublePuppetUtil
 
 	AS               *appservice.AppService
 	EventProcessor   *appservice.EventProcessor
@@ -184,20 +206,31 @@ type Bridge struct {
 	Child ChildOverride
 
 	manualStop chan int
+	Stopping   bool
+
+	latestState *status.BridgeState
+
+	Websocket                      bool
+	wsStopPinger                   chan struct{}
+	wsStarted                      chan struct{}
+	wsStopped                      chan struct{}
+	wsShortCircuitReconnectBackoff chan struct{}
+	wsStartupWait                  *sync.WaitGroup
 }
 
 type Crypto interface {
-	HandleMemberEvent(*event.Event)
-	Decrypt(*event.Event) (*event.Event, error)
-	Encrypt(id.RoomID, event.Type, *event.Content) error
-	WaitForSession(id.RoomID, id.SenderKey, id.SessionID, time.Duration) bool
-	RequestSession(id.RoomID, id.SenderKey, id.SessionID, id.UserID, id.DeviceID)
-	ResetSession(id.RoomID)
-	Init() error
+	HandleMemberEvent(context.Context, *event.Event)
+	Decrypt(context.Context, *event.Event) (*event.Event, error)
+	Encrypt(context.Context, id.RoomID, event.Type, *event.Content) error
+	WaitForSession(context.Context, id.RoomID, id.SenderKey, id.SessionID, time.Duration) bool
+	RequestSession(context.Context, id.RoomID, id.SenderKey, id.SessionID, id.UserID, id.DeviceID)
+	ResetSession(context.Context, id.RoomID)
+	Init(ctx context.Context) error
 	Start()
 	Stop()
-	Reset(startAfterReset bool)
+	Reset(ctx context.Context, startAfterReset bool)
 	Client() *mautrix.Client
+	ShareKeys(context.Context) error
 }
 
 func (br *Bridge) GenerateRegistration() {
@@ -230,6 +263,7 @@ func (br *Bridge) GenerateRegistration() {
 }
 
 func (br *Bridge) InitVersion(tag, commit, buildTime string) {
+	br.baseVersion = br.Version
 	if len(tag) > 0 && tag[0] == 'v' {
 		tag = tag[1:]
 	}
@@ -253,90 +287,112 @@ func (br *Bridge) InitVersion(tag, commit, buildTime string) {
 	}
 	mautrix.DefaultUserAgent = fmt.Sprintf("%s/%s %s", br.Name, br.Version, mautrix.DefaultUserAgent)
 	br.VersionDesc = fmt.Sprintf("%s %s (%s with %s)", br.Name, br.Version, buildTime, runtime.Version())
+	br.commit = commit
 	br.BuildTime = buildTime
 }
 
 var MinSpecVersion = mautrix.SpecV11
 
-func (br *Bridge) ensureConnection() {
+func (br *Bridge) ensureConnection(ctx context.Context) {
 	for {
-		versions, err := br.Bot.Versions()
+		versions, err := br.Bot.Versions(ctx)
 		if err != nil {
 			br.ZLog.Err(err).Msg("Failed to connect to homeserver, retrying in 10 seconds...")
 			time.Sleep(10 * time.Second)
-			continue
-		}
-		br.SpecVersions = *versions
-		if br.Config.Homeserver.Software == bridgeconfig.SoftwareHungry && !versions.UnstableFeatures["com.beeper.hungry"] {
-			br.ZLog.WithLevel(zerolog.FatalLevel).Msg("The config claims the homeserver is hungryserv, but the /versions response didn't confirm it")
-			os.Exit(18)
-		} else if !versions.ContainsGreaterOrEqual(MinSpecVersion) {
-			br.ZLog.WithLevel(zerolog.FatalLevel).
-				Stringer("server_supports", versions.GetLatest()).
-				Stringer("bridge_requires", MinSpecVersion).
-				Msg("The homeserver is outdated (supported spec versions are below minimum required by bridge)")
-			os.Exit(18)
-		} else if fr, ok := br.Child.(CSFeatureRequirer); ok {
-			if msg, hasFeatures := fr.CheckFeatures(versions); !hasFeatures {
-				br.ZLog.WithLevel(zerolog.FatalLevel).Msg(msg)
-				os.Exit(18)
-			}
-		}
-
-		resp, err := br.Bot.Whoami()
-		if err != nil {
-			if errors.Is(err, mautrix.MUnknownToken) {
-				br.ZLog.WithLevel(zerolog.FatalLevel).Msg("The as_token was not accepted. Is the registration file installed in your homeserver correctly?")
-				os.Exit(16)
-			} else if errors.Is(err, mautrix.MExclusive) {
-				br.ZLog.WithLevel(zerolog.FatalLevel).Msg("The as_token was accepted, but the /register request was not. Are the homeserver domain and username template in the config correct, and do they match the values in the registration?")
-				os.Exit(16)
-			}
-			br.ZLog.Err(err).Msg("Failed to connect to homeserver, retrying in 10 seconds...")
-			time.Sleep(10 * time.Second)
-			continue
-		} else if resp.UserID != br.Bot.UserID {
-			br.ZLog.WithLevel(zerolog.FatalLevel).
-				Stringer("got_user_id", resp.UserID).
-				Stringer("expected_user_id", br.Bot.UserID).
-				Msg("Unexpected user ID in whoami call")
-			os.Exit(17)
-		}
-
-		if br.SpecVersions.UnstableFeatures["fi.mau.msc2659"] && br.AS.Host.IsConfigured() {
-			txnID := br.Bot.TxnID()
-			resp, err := br.Bot.AppservicePing(br.Config.AppService.ID, txnID)
-			if err != nil {
-				evt := br.ZLog.WithLevel(zerolog.FatalLevel).Err(err).Str("txn_id", txnID)
-				var httpErr mautrix.HTTPError
-				if errors.As(err, &httpErr) && httpErr.RespError != nil {
-					if val, ok := httpErr.RespError.ExtraData["body"].(string); ok {
-						val = strings.TrimSpace(val)
-						valBytes := []byte(val)
-						if json.Valid(valBytes) {
-							evt.RawJSON("body", valBytes)
-						} else {
-							evt.Str("body", val)
-						}
-					}
-				}
-				evt.Msg("Homeserver -> bridge connection is not working")
-				os.Exit(13)
-			}
-			br.ZLog.Debug().
-				Str("txn_id", txnID).
-				Int64("duration_ms", resp.DurationMS).
-				Msg("Homeserver -> bridge connection works")
 		} else {
-			br.ZLog.Debug().Msg("Homeserver does not support checking status of homeserver -> bridge connection")
+			br.SpecVersions = *versions
+			break
 		}
-
-		break
 	}
+
+	if br.Config.Homeserver.Software == bridgeconfig.SoftwareHungry && !br.SpecVersions.Supports(mautrix.BeeperFeatureHungry) {
+		br.ZLog.WithLevel(zerolog.FatalLevel).Msg("The config claims the homeserver is hungryserv, but the /versions response didn't confirm it")
+		os.Exit(18)
+	} else if !br.SpecVersions.ContainsGreaterOrEqual(MinSpecVersion) {
+		br.ZLog.WithLevel(zerolog.FatalLevel).
+			Stringer("server_supports", br.SpecVersions.GetLatest()).
+			Stringer("bridge_requires", MinSpecVersion).
+			Msg("The homeserver is outdated (supported spec versions are below minimum required by bridge)")
+		os.Exit(18)
+	} else if fr, ok := br.Child.(CSFeatureRequirer); ok {
+		if msg, hasFeatures := fr.CheckFeatures(&br.SpecVersions); !hasFeatures {
+			br.ZLog.WithLevel(zerolog.FatalLevel).Msg(msg)
+			os.Exit(18)
+		}
+	}
+
+	resp, err := br.Bot.Whoami(ctx)
+	if err != nil {
+		if errors.Is(err, mautrix.MUnknownToken) {
+			br.ZLog.WithLevel(zerolog.FatalLevel).Msg("The as_token was not accepted. Is the registration file installed in your homeserver correctly?")
+		} else if errors.Is(err, mautrix.MExclusive) {
+			br.ZLog.WithLevel(zerolog.FatalLevel).Msg("The as_token was accepted, but the /register request was not. Are the homeserver domain, bot username and username template in the config correct, and do they match the values in the registration?")
+		} else {
+			br.ZLog.WithLevel(zerolog.FatalLevel).Err(err).Msg("/whoami request failed with unknown error")
+		}
+		os.Exit(16)
+	} else if resp.UserID != br.Bot.UserID {
+		br.ZLog.WithLevel(zerolog.FatalLevel).
+			Stringer("got_user_id", resp.UserID).
+			Stringer("expected_user_id", br.Bot.UserID).
+			Msg("Unexpected user ID in whoami call")
+		os.Exit(17)
+	}
+
+	if br.Websocket {
+		br.ZLog.Debug().Msg("Websocket mode: no need to check status of homeserver -> bridge connection")
+		return
+	} else if !br.SpecVersions.Supports(mautrix.FeatureAppservicePing) {
+		br.ZLog.Debug().Msg("Homeserver does not support checking status of homeserver -> bridge connection")
+		return
+	}
+	var pingResp *mautrix.RespAppservicePing
+	var txnID string
+	var retryCount int
+	const maxRetries = 6
+	for {
+		txnID = br.Bot.TxnID()
+		pingResp, err = br.Bot.AppservicePing(ctx, br.Config.AppService.ID, txnID)
+		if err == nil {
+			break
+		}
+		var httpErr mautrix.HTTPError
+		var pingErrBody string
+		if errors.As(err, &httpErr) && httpErr.RespError != nil {
+			if val, ok := httpErr.RespError.ExtraData["body"].(string); ok {
+				pingErrBody = strings.TrimSpace(val)
+			}
+		}
+		outOfRetries := retryCount >= maxRetries
+		level := zerolog.ErrorLevel
+		if outOfRetries {
+			level = zerolog.FatalLevel
+		}
+		evt := br.ZLog.WithLevel(level).Err(err).Str("txn_id", txnID)
+		if pingErrBody != "" {
+			bodyBytes := []byte(pingErrBody)
+			if json.Valid(bodyBytes) {
+				evt.RawJSON("body", bodyBytes)
+			} else {
+				evt.Str("body", pingErrBody)
+			}
+		}
+		if outOfRetries {
+			evt.Msg("Homeserver -> bridge connection is not working")
+			os.Exit(13)
+		}
+		evt.Msg("Homeserver -> bridge connection is not working, retrying in 5 seconds...")
+		time.Sleep(5 * time.Second)
+		retryCount++
+	}
+	br.ZLog.Debug().
+		Str("txn_id", txnID).
+		Int64("duration_ms", pingResp.DurationMS).
+		Msg("Homeserver -> bridge connection works")
 }
 
-func (br *Bridge) fetchMediaConfig() {
-	cfg, err := br.Bot.GetMediaConfig()
+func (br *Bridge) fetchMediaConfig(ctx context.Context) {
+	cfg, err := br.Bot.GetMediaConfig(ctx)
 	if err != nil {
 		br.ZLog.Warn().Err(err).Msg("Failed to fetch media config")
 	} else {
@@ -344,37 +400,36 @@ func (br *Bridge) fetchMediaConfig() {
 	}
 }
 
-func (br *Bridge) UpdateBotProfile() {
+func (br *Bridge) UpdateBotProfile(ctx context.Context) {
 	br.ZLog.Debug().Msg("Updating bot profile")
 	botConfig := &br.Config.AppService.Bot
 
 	var err error
 	var mxc id.ContentURI
 	if botConfig.Avatar == "remove" {
-		err = br.Bot.SetAvatarURL(mxc)
+		err = br.Bot.SetAvatarURL(ctx, mxc)
 	} else if !botConfig.ParsedAvatar.IsEmpty() {
-		err = br.Bot.SetAvatarURL(botConfig.ParsedAvatar)
+		err = br.Bot.SetAvatarURL(ctx, botConfig.ParsedAvatar)
 	}
 	if err != nil {
 		br.ZLog.Warn().Err(err).Msg("Failed to update bot avatar")
 	}
 
 	if botConfig.Displayname == "remove" {
-		err = br.Bot.SetDisplayName("")
+		err = br.Bot.SetDisplayName(ctx, "")
 	} else if len(botConfig.Displayname) > 0 {
-		err = br.Bot.SetDisplayName(botConfig.Displayname)
+		err = br.Bot.SetDisplayName(ctx, botConfig.Displayname)
 	}
 	if err != nil {
 		br.ZLog.Warn().Err(err).Msg("Failed to update bot displayname")
 	}
 
-	if br.Config.Homeserver.Software == bridgeconfig.SoftwareHungry && br.BeeperNetworkName != "" {
+	if br.SpecVersions.Supports(mautrix.BeeperFeatureArbitraryProfileMeta) && br.BeeperNetworkName != "" {
 		br.ZLog.Debug().Msg("Setting contact info on the appservice bot")
-		br.Bot.BeeperUpdateProfile(map[string]any{
+		br.Bot.BeeperUpdateProfile(ctx, map[string]any{
 			"com.beeper.bridge.service":       br.BeeperServiceName,
 			"com.beeper.bridge.network":       br.BeeperNetworkName,
 			"com.beeper.bridge.is_bridge_bot": true,
-			"com.beeper.bridge.is_bot":        true,
 		})
 	}
 }
@@ -419,7 +474,15 @@ func (br *Bridge) validateConfig() error {
 	case br.Config.AppService.Database.URI == "postgres://user:password@host/database?sslmode=disable":
 		return errors.New("appservice.database not configured")
 	default:
-		return br.Config.Bridge.Validate()
+		err := br.Config.Bridge.Validate()
+		if err != nil {
+			return err
+		}
+		validator, ok := br.Child.(ConfigValidatingBridge)
+		if ok {
+			return validator.ValidateConfig()
+		}
+		return nil
 	}
 }
 
@@ -455,13 +518,12 @@ func (br *Bridge) init() {
 	}
 	defaultCtxLog := br.ZLog.With().Bool("default_context_log", true).Caller().Logger()
 	zerolog.TimeFieldFormat = time.RFC3339Nano
+	zerolog.CallerMarshalFunc = exzerolog.CallerWithFunctionName
 	zerolog.DefaultContextLogger = &defaultCtxLog
+	deflog.Logger = br.ZLog.With().Bool("global_log", true).Caller().Logger()
 	br.Log = maulogadapt.ZeroAsMau(br.ZLog)
 
-	br.AS = br.Config.MakeAppService()
-	br.AS.DoublePuppetValue = br.Name
-	br.AS.GetProfile = br.getProfile
-	br.AS.Log = *br.ZLog
+	br.DoublePuppet = &doublePuppetUtil{br: br, log: br.ZLog.With().Str("component", "double puppet").Logger()}
 
 	err = br.validateConfig()
 	if err != nil {
@@ -469,7 +531,6 @@ func (br *Bridge) init() {
 		os.Exit(11)
 	}
 
-	br.Bot = br.AS.BotIntent()
 	br.ZLog.Info().
 		Str("name", br.Name).
 		Str("version", br.Version).
@@ -505,7 +566,13 @@ func (br *Bridge) init() {
 
 	br.ZLog.Debug().Msg("Initializing state store")
 	br.StateStore = sqlstatestore.NewSQLStateStore(br.DB, dbutil.ZeroLogger(br.ZLog.With().Str("db_section", "matrix_state").Logger()), true)
+
+	br.AS = br.Config.MakeAppService()
+	br.AS.DoublePuppetValue = br.Name
+	br.AS.GetProfile = br.getProfile
+	br.AS.Log = *br.ZLog
 	br.AS.StateStore = br.StateStore
+	br.Bot = br.AS.BotIntent()
 
 	br.ZLog.Debug().Msg("Initializing Matrix event processor")
 	br.EventProcessor = appservice.NewEventProcessor(br.AS)
@@ -532,11 +599,50 @@ func (br *Bridge) init() {
 	br.Child.Init()
 }
 
+type zerologPQError pq.Error
+
+func (zpe *zerologPQError) MarshalZerologObject(evt *zerolog.Event) {
+	maybeStr := func(field, value string) {
+		if value != "" {
+			evt.Str(field, value)
+		}
+	}
+	maybeStr("severity", zpe.Severity)
+	if name := zpe.Code.Name(); name != "" {
+		evt.Str("code", name)
+	} else if zpe.Code != "" {
+		evt.Str("code", string(zpe.Code))
+	}
+	//maybeStr("message", zpe.Message)
+	maybeStr("detail", zpe.Detail)
+	maybeStr("hint", zpe.Hint)
+	maybeStr("position", zpe.Position)
+	maybeStr("internal_position", zpe.InternalPosition)
+	maybeStr("internal_query", zpe.InternalQuery)
+	maybeStr("where", zpe.Where)
+	maybeStr("schema", zpe.Schema)
+	maybeStr("table", zpe.Table)
+	maybeStr("column", zpe.Column)
+	maybeStr("data_type_name", zpe.DataTypeName)
+	maybeStr("constraint", zpe.Constraint)
+	maybeStr("file", zpe.File)
+	maybeStr("line", zpe.Line)
+	maybeStr("routine", zpe.Routine)
+}
+
 func (br *Bridge) LogDBUpgradeErrorAndExit(name string, err error) {
-	br.ZLog.WithLevel(zerolog.FatalLevel).
+	logEvt := br.ZLog.WithLevel(zerolog.FatalLevel).
 		Err(err).
-		Str("db_section", name).
-		Msg("Failed to initialize database")
+		Str("db_section", name)
+	var errWithLine *dbutil.PQErrorWithLine
+	if errors.As(err, &errWithLine) {
+		logEvt.Str("sql_line", errWithLine.Line)
+	}
+	var pqe *pq.Error
+	if errors.As(err, &pqe) {
+		logEvt.Object("pq_error", (*zerologPQError)(pqe))
+	}
+	logEvt.Msg("Failed to initialize database")
 	if sqlError := (&sqlite3.Error{}); errors.As(err, sqlError) && sqlError.Code == sqlite3.ErrCorrupt {
 		os.Exit(18)
 	} else if errors.Is(err, dbutil.ErrForeignTables) {
@@ -549,27 +655,44 @@ func (br *Bridge) LogDBUpgradeErrorAndExit(name string, err error) {
 	os.Exit(15)
 }
 
+func (br *Bridge) WaitWebsocketConnected() {
+	if br.wsStartupWait != nil {
+		br.wsStartupWait.Wait()
+	}
+}
+
 func (br *Bridge) start() {
 	br.ZLog.Debug().Msg("Running database upgrades")
-	err := br.DB.Upgrade()
+	err := br.DB.Upgrade(br.ZLog.With().Str("db_section", "main").Logger().WithContext(context.TODO()))
 	if err != nil {
 		br.LogDBUpgradeErrorAndExit("main", err)
-	} else if err = br.StateStore.Upgrade(); err != nil {
+	} else if err = br.StateStore.Upgrade(br.ZLog.With().Str("db_section", "matrix_state").Logger().WithContext(context.TODO())); err != nil {
 		br.LogDBUpgradeErrorAndExit("matrix_state", err)
 	}
 
-	if br.AS.Host.IsConfigured() {
+	if br.Config.Homeserver.Websocket || len(br.Config.Homeserver.WSProxy) > 0 {
+		br.Websocket = true
+		br.ZLog.Debug().Msg("Starting application service websocket")
+		var wg sync.WaitGroup
+		wg.Add(1)
+		br.wsStartupWait = &wg
+		br.wsShortCircuitReconnectBackoff = make(chan struct{})
+		go br.startWebsocket(&wg)
+	} else if br.AS.Host.IsConfigured() {
 		br.ZLog.Debug().Msg("Starting application service HTTP server")
 		go br.AS.Start()
 	} else {
-		br.ZLog.Debug().Msg("Appservice config doesn't have port nor unix socket path, not starting HTTP server")
+		br.ZLog.WithLevel(zerolog.FatalLevel).Msg("Neither appservice HTTP listener nor websocket is enabled")
+		os.Exit(23)
 	}
 	br.ZLog.Debug().Msg("Checking connection to homeserver")
-	br.ensureConnection()
-	go br.fetchMediaConfig()
+
+	ctx := br.ZLog.WithContext(context.Background())
+	br.ensureConnection(ctx)
+	go br.fetchMediaConfig(ctx)
 
 	if br.Crypto != nil {
-		err = br.Crypto.Init()
+		err = br.Crypto.Init(ctx)
 		if err != nil {
 			br.ZLog.WithLevel(zerolog.FatalLevel).Err(err).Msg("Error initializing end-to-bridge encryption")
 			os.Exit(19)
@@ -577,18 +700,23 @@ func (br *Bridge) start() {
 	}
 
 	br.ZLog.Debug().Msg("Starting event processor")
-	br.EventProcessor.Start()
+	br.EventProcessor.Start(ctx)
 
-	go br.UpdateBotProfile()
+	go br.UpdateBotProfile(ctx)
 	if br.Crypto != nil {
 		go br.Crypto.Start()
 	}
 
 	br.Child.Start()
+	br.WaitWebsocketConnected()
 	br.AS.Ready = true
 
 	if br.Config.Bridge.GetResendBridgeInfo() {
 		go br.ResendBridgeInfo()
+	}
+	if br.Websocket && br.Config.Homeserver.WSPingInterval > 0 {
+		br.wsStopPinger = make(chan struct{}, 1)
+		go br.websocketServerPinger()
 	}
 }
 
@@ -605,21 +733,46 @@ func (br *Bridge) ResendBridgeInfo() {
 	}
 	br.ZLog.Info().Msg("Re-sending bridge info state event to all portals")
 	for _, portal := range br.Child.GetAllIPortals() {
-		portal.UpdateBridgeInfo()
+		portal.UpdateBridgeInfo(context.TODO())
 	}
 	br.ZLog.Info().Msg("Finished re-sending bridge info state events")
 }
 
+func sendStopSignal(ch chan struct{}) {
+	if ch != nil {
+		select {
+		case ch <- struct{}{}:
+		default:
+		}
+	}
+}
+
 func (br *Bridge) stop() {
+	br.Stopping = true
 	if br.Crypto != nil {
 		br.Crypto.Stop()
 	}
+	waitForWS := false
+	if br.AS.StopWebsocket != nil {
+		br.ZLog.Debug().Msg("Stopping application service websocket")
+		br.AS.StopWebsocket(appservice.ErrWebsocketManualStop)
+		waitForWS = true
+	}
 	br.AS.Stop()
+	sendStopSignal(br.wsStopPinger)
+	sendStopSignal(br.wsShortCircuitReconnectBackoff)
 	br.EventProcessor.Stop()
 	br.Child.Stop()
-	err := br.DB.RawDB.Close()
+	err := br.DB.Close()
 	if err != nil {
 		br.ZLog.Warn().Err(err).Msg("Error closing database")
+	}
+	if waitForWS {
+		select {
+		case <-br.wsStopped:
+		case <-time.After(4 * time.Second):
+			br.ZLog.Warn().Msg("Timed out waiting for websocket to close")
+		}
 	}
 }
 
@@ -628,6 +781,25 @@ func (br *Bridge) ManualStop(exitCode int) {
 		br.manualStop <- exitCode
 	} else {
 		os.Exit(exitCode)
+	}
+}
+
+type VersionJSONOutput struct {
+	Name string
+	URL  string
+
+	Version          string
+	IsRelease        bool
+	Commit           string
+	FormattedVersion string
+	BuildTime        string
+
+	OS   string
+	Arch string
+
+	Mautrix struct {
+		Version string
+		Commit  string
 	}
 }
 
@@ -648,6 +820,24 @@ func (br *Bridge) Main() {
 		os.Exit(0)
 	} else if *version {
 		fmt.Println(br.VersionDesc)
+		return
+	} else if *versionJSON {
+		output := VersionJSONOutput{
+			URL:  br.URL,
+			Name: br.Name,
+
+			Version:          br.baseVersion,
+			IsRelease:        br.Version == br.baseVersion,
+			Commit:           br.commit,
+			FormattedVersion: br.Version,
+			BuildTime:        br.BuildTime,
+
+			OS:   runtime.GOOS,
+			Arch: runtime.GOARCH,
+		}
+		output.Mautrix.Commit = mautrix.Commit
+		output.Mautrix.Version = mautrix.Version
+		_ = json.NewEncoder(os.Stdout).Encode(output)
 		return
 	} else if flagHandler, ok := br.Child.(FlagHandlingBridge); ok && flagHandler.HandleFlags() {
 		return